Everything seems to be working perfectly
Added functionality: on network status changed, the bot restarts the timer and checks whether to post and stuff

LOG_DATA compiler conditional removed. Popup message on image posted was disabled. Reason: surprisingly, everything seems to be working fine and the program is yet to throw a single exception. Time to leave it alone running on the background of my computer.